Remove unnecessary helper
